backend/Code-Essence/backend/requirements.txt
flask
boto3
langchain
langchain-community
gitpython
langchain-aws
redis
transformers
Authlib
